% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinMS.R
\name{binMS}
\alias{binMS}
\title{Consolidate mass spectrometry observations}
\usage{
binMS(mass_spec, mtoz, charge, mass, time_peak_reten, ms_inten, time_pr_range,
  mass_range, charge_range, mtoz_diff, time_pr_diff, ...)
}
\arguments{
\item{mass_spec}{}

\item{mtoz}{}

\item{charge}{}

\item{mass}{}

\item{time_peak_reten}{}

\item{ms_inten}{}

\item{time_pr_range}{}

\item{mass_range}{}

\item{charge_range}{}

\item{mtoz_diff}{}

\item{time_pr_diff}{}

\item{...}{}
}
\value{
out
}
\description{
Combines mass spectrometry observations that are believed to belong to the
same true observation into a single observation.  In concept, the data
produced by the mass spectrometer may contain multiple reads for a single
underlying compound.  Thus, this function attempts to recover these
underlying compounds through a binning procedure, described in more detail
in \code{Details}.
}
\author{
Pritchard
}

