% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioact}
\alias{bioact}
\title{Bioactivity}
\format{A \code{list} containing relative growth inhibition bioactivity
    levels for the bacteria and virus strains listed below.  Each of the
    following elements in the \code{list} is a \code{data.frame} with 3 rows
    and 44 columns (with the exception of fg ********* which has 2 rows).
    The rows in each \code{data.frame} correspond to replications of the data
    collection process, while the columns correspond to relative growth
    inhibition bioactivity levels when subject to peptide libraries across
    fractions 1-43 and fraction 47.

\describe{

    \item{ec}{E. Coli}

    \item{bc}{S. aureus}

    \item{pc}{K. pneumoniae}

    \item{oc}{A. baumannii}

    \item{ef}{E. cloacae}

    \item{ab}{??}

    \item{pa}{??}

    \item{fg}{??}

}}
\usage{
data(bioact)
}
\description{
The mass spectrometry data collected for and described in Kirkpatrick et
al. (2016).  See the paper for a full description of the data collection
process, or the package vignette for an abridged description.
}
\keyword{datasets}

