% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MsDat.R
\name{checkValInp_msDat}
\alias{checkValInp_msDat}
\title{Check for valid msDat arguments}
\usage{
checkValInp_msDat(mass_spec, mtoz, charge)
}
\arguments{
\item{mass_spec}{Either a \code{matrix} or \code{data.frame}.  This object
  contains mass-to-charge values, charge information, and intensity data.
  The mass spectrometry data is expected to be in a form such that a given
  column provides the mass spectrometry intensity values for a particular
  fraction.  Then either 0, 1, or 2 additional columns may be included in the
  parameter input containing a possible column for the mass-to-charge values
  and a possible column for the charge information. Any ordering of the
  columns is allowed.

  For example, suppose that a collection of mass spectrometry readings has
  provided data for 50 fractions across 20,000 mass-to-charge values.  Then
  the input for \code{mass_spec} should be a \code{matrix} or
  \code{data.frame} with 20,000 rows and one of 50, 51, or 52 columns.  Each
  row in the 50 columns containing the mass spectrometry readings should be
  for the same mass-to-charge value and charge information.  If columns are
  included for either the mass-to-charge values or charge information then
  the entries in these columns should provide the identifying information for
  the mass spectrometry data in the same row.}

\item{mtoz}{A vector of either length 1 or length equal to the number of
  mass-to-charge values for which mass spectrometry data was collected, and
  which helps identify the mass-to-charge values for this data in one of
  several ways.

  One way to provide the information is to provide a vector where each entry
  provides the mass-to-charge value for a corresponding row of mass
  spectrometry data.  Then the \code{k}-th entry of the vector would provide
  the mass-to-charge value for the \code{k}-th row of the mass spectrometry
  data.

  A second way is to provide a single number which provides the column number
  in the input to the \code{mass_spec} parameter for a column which contains
  this information.

  A third way is provide a single character string which provides the column
  name in the input to the \code{mass_spec} parameter for a column which
  contains this information.  Partial matching is allowed.}

\item{charge}{The information for the \code{charge} parameter can be provided
  in the same manner as for the mass-to-charge values.}
}
\description{
Check that arguments \code{mass_spec}, \code{mtoz}, and \code{charge} are
of the right data type
}

