% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MsDat-API.R
\name{dimnames.msDat}
\alias{dimnames.msDat}
\title{Mass spectrometry abundances fraction names}
\usage{
\method{dimnames}{msDat}(msObj)
}
\arguments{
\item{msObj}{An object of class \code{binMS}, \code{filterMS}, or
  \code{msDat}}

\item{value}{Either \code{NULL} or a character vector or an object that can
  be coerced to a character vector.  If non-\code{NULL}, then must be of the same
  length as the number of fractions in the mass spectrometry data.}
}
\description{
Retrieve or set the mass spectrometry abundances fraction names for objects
of class \code{binMS}, \code{filterMS}, or \code{msDat}.
}

