% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankLasso-First_Level_Fcns.R
\name{conv_bioact}
\alias{conv_bioact}
\title{Convert bioactivity data}
\usage{
conv_bioact(bioMat, bioIdx, useAve)
}
\arguments{
\item{bioMat}{The object returned by the format_bio function}

\item{bioIdx}{An integer vector providing the indices (and hence fractions of
interest) for the bioactivity data which are to be used in the Lasso model}
}
\description{
Creates the response vector for use by the \code{lars} function. This
involves transposing the data and reducing the data to the fractions of
interest.  If replicates are to be used, then the fractions containing the
replicates are concatenated together.  If the replicates are to be averaged,
then the data within a fraction is averaged before concatenating the values.
}
\details{
It is important to define how the ordering of the replicates is performed so
that this behavior matches that of \code{\link{conv_ms}}.  The current
implementation places replicates within a fraction consecutively, so that for
example, the data may look like: fraction 1 replicate 1, fraction 1 replicate
2, fraction 1 replicate 3, fraction 2 replicate 1, fraction 2 replicate 2,
fraction 2 replicate 3, fraction 3 replicate 1, ...
}

