% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankLasso.R
\name{conv_ms}
\alias{conv_ms}
\title{Convert mass spectrometry data}
\usage{
conv_ms(ms, msIdx, nRepl, useAve)
}
\arguments{
\item{ms}{A \code{matrix} containing mass spectrometry intensity readings.
  Each column provides the mass spectrometry values for a given fraction, and
  each row provides the mass spectrometry values for a given mass-to-charge
  ratio value across the fractions.}

\item{msIdx}{An integer vector providing the indices (and hence fractions)
  for the mass spectrometry data which are to be used in the Lasso model}

\item{nRepl}{An integer giving the number of replications provided in the
  bioactivity data.  Note that this paramter is not needed when \code{useAve}
  is \code{TRUE}.}

\item{useAve}{A logical value specifying whether or not to average replicate
bioactivity observations.  Ignored if only one bioactivity observation is
provided.}
}
\value{
A matrix providing the mass spectrometry (and hence the explanetory)
  data for the Lasso model.  The rows are the observations (possibly expanded
  from the original data) for the model, and the columns are the compounds
  (and hence are the predictor variables).
}
\description{
Creates the explanetory data matrix for use by the \code{lars} function. This
involves transposing the data and reducing the data to the fractions of
interest.  Furthermore, if there are replicates for the bioactivity data,
then multiple copies of each fraction of interest are made, one for each
replicate.
}
\details{
It is important to define how the ordering of the replicates is performed so
that this behavior matches that of \code{\link{conv_bioact}}.  The current
implementation places replicates within a fraction consecutively, so that for
example, the data may look like (going down the rows): fraction 1,
fraction 1, fraction 1, fraction 2, fraction 2, fraction 2, fraction 3, ...
}

