% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FilterMS.R
\name{filterMS}
\alias{filterMS}
\title{Filter compounds from mass spectrometry data}
\usage{
filterMS(msObj, region, border = "all", bor_rat = 0.05, min_inten = 1000,
  max_chg = 7L)
}
\arguments{
\item{msObj}{An object of class \code{\link{msDat}}}

\item{region}{A vector either of mode numeric or character.  If numeric then
  the entries should provide the indices for the region of interest in the
  mass spectrometry data in the argument for \code{msObj}.  If character then
  the entries should uniquely specify the region of interest through partial
  string matching.}

\item{border}{Either a character string with the value \code{"all"}, or a
  single numeric value specifying the number of fractions to either side of
  the region of interest to comprise the bordering region}

\item{bor_rat}{A single numeric value between 0 and 1, inclusive}

\item{min_inten}{A single numeric value greater than or equal to 0}

\item{max_chg}{A single numeric value specifying the maximum charge which a
  compound may exhibit}
}
\value{
Returns an object of class of classes \code{link{msDat}} and
  \code{filterMS}.  This object is a \code{list} with elements described
  below.  The classes are equipped with a summary function.

  \describe{

  \item{\code{msObj}}{ An object of class \code{\link{msDat}} such that the
  encapsulated mass spectrometry data corresponds to each of the candidate
  compounds that satisfed each of the criteria.  If no criteria are satisfied
  then \code{NULL} is returned. }

  \item{\code{cmp_by_crit}}{ A list containing \code{data.frame}s, one for
  each criterion. Each row (if any) in one of the sub-\code{data.frame}s
  contains the mass-to-charge and charge information for a candidate compound
  that satisfies the criterion represented by the \code{data.frame}; all of
  the compounds that satisfied the criterion are included in the data.  The
  \code{data.frame}s are named \code{c1}, \code{c2}, etc corresponding to
  criterion 1, criterion 2, and so on. }

  \item{\code{summ_info}}{ A list containing information pertaining to the
  filtering process; for use by the summary function. }

  }
}
\description{
Filters mass spectrometry data using a set of criteria, described in
\code{Details}. Returns an object of classes \code{\link{msDat}} and
\code{filterMS}.
}
\details{
Attempts to filter out candidate compounds via subject-matter
  knowledge, with the goal of removing spurious noise from downstream models.
  The criteria for the downstream inclusion of a candidate compound is listed
  below.

  \enumerate{

  \item The m/z intensity maximum must fall inside the range of the
  bioactivity region of interest

  \item The ratio of the m/z intensity of a species in the areas bordering
  the region of interest and the species maximum intensity must be less than
  \code{bor_rat}

  \item The immediately right adjacent fraction to its maximum intensity
  fraction for a species must have a non-zero abundance

  \item Each fraction in the region of interest must have intensity greater
  than \code{min_inten}

  \item Compound charge state must be less than or equal to \code{max_chg}

  }
}

