% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankLasso-First_Level_Fcns.R
\name{getCmpIdx}
\alias{getCmpIdx}
\title{Extract compound ordering from \code{lars} fit}
\usage{
getCmpIdx(lars_fit)
}
\arguments{
\item{lars_fit}{An object of class \code{lars}}
}
\value{
An integer vector indexing (some of the) compounds by the order in
  which they first enter the Lasso model, from first to last.
}
\description{
Returns an integer vector indexing (some of the) compounds by the order in
which they first enter the Lasso model, from first to last.
}
\details{
A \code{lars} object is a list containing (among other things) an
  obect '\code{actions}'. This object is a list of all the actions taken in
  the model Lasso path, where actions are taken to be either adding or
  removing a predictor variable from model.  If the variable is added to the
  model, then the column index is inserted into the next position in the
  list.  If the variable is removed from the model, then -1 times the column
  index of the variable is inserted into the next position in the list.

  \code{getCmpIdx} works by reducing '\code{actions}' list to the first
  positive time that a number is in the list.
}

