% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankLasso-First_Level_Fcns.R
\name{getRegionIdx}
\alias{getRegionIdx}
\title{Calculates indices for the fractions of interest}
\usage{
getRegionIdx(msDat, bioact, region)
}
\arguments{
\item{msDat}{An object of class \code{\link{msDat}} containing the mass
  spectrometry data and identifying information}

\item{bioact}{Either a non-list vector, a matrix, or a data frame providing
  bioactivity data.  If a non-list vector, then it is assumed that each entry
  corresponds to a particular fraction.  If the data is 2-dimensional, then
  it is assumed that each column corresponds to a particular fraction, and
  that each row corresponds to a particular bioactivity replicate.}

\item{region}{Either \code{NULL}, a non-list vector, a matrix, or a list
  containing exactly two atomic vectors, providing information specifying
  which fractions are to be included in the Lasso model.  Note that a data
  frame satisfies these requirements.

  If \code{NULL}, then it is assumed that all fractions included in the data
  are to be used in the model.  This requires that the number of fractions in
  the data for the parameter passed to \code{bioact} be the same as the
  number of fractions in the mass spectrometry data for the paramter passed
  to \code{msDat}.  It further assumes that the \code{k}-th column must refer
  to the same fraction for both the mass spectrometry data and the
  bioactivity data for every \code{k}.

  If a non-list vector then it must be either a numeric or charcter vector,
  such that the vector specifies which columns (and hence which fractions) to
  include in the model.  If the vector is numeric, then the desired columns
  are specified by number, and if the vector is character, then the desired
  columns are specified by name (partial matching is allowed).  Note that
  this assumes that the corresponding columns in the mass spectrometry data
  and the bioactivity data refer to the same fractions.

  If a matrix then it must be either numeric or character with exactly two
  columns - one is to be named \code{ms} and the other is to be named
  \code{bio}.  The \code{ms} column specifies the columns (and hence
  fractions) to include in the model from the mass spectrometry data, either
  as a vector of the column numbers or the column names.  The \code{bio}
  vector specifies the columns (and hence fractions) to include in the model
  from the bioactivity data, either as a vector of the column numbers or the
  column names.  It is assumed that two entries from a given row refer to the
  same fraction.

  If a list, then it must be a list with two named non-list vectors of equal
  length - one is to be named \code{ms} and the other is to be named
  \code{bio} (note that a \code{n x 2} data frame satisfies this
  requirement).  The \code{ms} vector specifies the columns (and hence
  fractions) to include in the model from the mass spectrometry data, either
  as a vector of the column numbers or the column names.  The \code{bio}
  vector specifies the columns (and hence fractions) to include in the model
  from the bioactivity data, either as a vector of the column numbers or the
  column names.  It is assumed that the column from the mass spectrometry
  data specified by the \code{k}-th value in the \code{ms} vector corresponds
  to the same fraction as the column specified by the \code{k}-th value in
  the \code{bio} vector, for each \code{k}.}
}
\value{
A list providing the indices for the fractions of interest
  for the mass spectrometry and bioactivity data.  The list contains
  the elements described below.  Note that the \code{k}-th index
  for the mass spectrometry index corresponds to the \code{k}-th
  index for the bioactivity index.

  \describe{

  \item{\code{ms}}{An integer vector providing the indices for the
  fractions of interest in the mass spectrometry data }

  \item{\code{bio}}{An integer vector providing the indices for the
  fractions of interest in the bioactivity data }

  }
}
\description{
See description of the \code{region} parameter for the behavior of
this function
}

