% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MsDat.R
\name{msDat}
\alias{msDat}
\title{Constructor for class \code{msDat}}
\usage{
msDat(mass_spec, mtoz, charge, ms_inten = NULL)
}
\arguments{
\item{mass_spec}{Either a \code{matrix} or \code{data.frame}.  This object
  must contain mass spectrometry abundances, and may optionally contain
  mass-to-charge values and/or charge state information.  The mass
  spectrometry data is expected to be in a form such that a given column
  provides the mass spectrometry intensity values for a particular fraction.
  Then either 0, 1, or 2 additional columns may be included in the parameter
  input containing a possible column for the mass-to-charge values and a
  possible column for the charge information. Any ordering of the columns is
  allowed.

  For example, suppose that a collection of mass spectrometry readings has
  provided data for 50 fractions across 20,000 mass-to-charge values.  Then
  the input for \code{mass_spec} should be a \code{matrix} or
  \code{data.frame} with 20,000 rows and one of 50, 51, or 52 columns.  Each
  row in the 50 columns containing the mass spectrometry readings should be
  for the same mass-to-charge value and charge information.  If columns are
  included for either the mass-to-charge values or charge information then
  the entries in these columns should provide the identifying information for
  the mass spectrometry data in the same row.}

\item{mtoz}{A vector of either length 1 or length equal to the number of
  mass-to-charge values for which mass spectrometry data was collected, and
  which helps identify the mass-to-charge values for this data in one of
  several ways.

  One way to provide the information is to provide a vector where each entry
  provides the mass-to-charge value for a corresponding row of mass
  spectrometry data.  Then the \code{k}-th entry of the vector would provide
  the mass-to-charge value for the \code{k}-th row of the mass spectrometry
  data.

  A second way is to provide a single number which provides the column number
  in the input to the \code{mass_spec} parameter for a column which contains
  this information.

  A third way is provide a single character string which provides the column
  name in the input to the \code{mass_spec} parameter for a column which
  contains this information.  Partial matching is allowed.}

\item{charge}{The information for the \code{charge} parameter can be provided
  in the same manner as for the mass-to-charge values.}
}
\value{
Returns an object of class \code{msDat}.  This class is a \code{list}
  with elements described below.  The class is equipped with a summary
  function.

  \describe{

  \item{\code{ms}}{ A \code{matrix} containing mass spectrometry intensity
  readings. Each column provides the mass spectrometry values for a given
  fraction, and each row provides the mass spectrometry values for a given
  mass-to-charge ratio value across the fractions. }

  \item{\code{mtoz}}{ A vector with length equal to the number of
  mass-to-charge values provided in the mass spectrometry data, such that the
  \code{k}-th entry in the vector provides the mass-to-charge value for the
  \code{k}-th row of mass spectrometry data }

  \item{\code{chg}}{ A vector with length equal to the number of
  mass-to-charge values provided in the mass spectrometry data, such that the
  \code{k}-th entry in the vector provides the charge information for the
  \code{k}-th row of mass spectrometry data }

  }
}
\description{
Creates a data structure encapsulating the mass spectrometry intensity
readings as well as identifying information
}

