% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinMS.R
\name{print.binMS}
\alias{print.binMS}
\title{Print routine for class \code{binMS}}
\usage{
\method{print}{binMS}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{binMS}}}

\item{...}{Arguments passed to dot-dot-dot are ignored}
}
\description{
Prints the number of m/z levels and fractions of the resultant mass spectrometry data
}

