% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FilterMS.R
\name{print.filterMS}
\alias{print.filterMS}
\title{Basic information for class \code{filterMS}}
\usage{
\method{print}{filterMS}(filtObj)
}
\description{
Displays the number of candidate compounds left in the data after filtering
}

