% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsDat.R
\name{print.msDat}
\alias{print.msDat}
\title{Print method for class \code{msDat}}
\usage{
\method{print}{msDat}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{msDat}}}

\item{...}{Arguments passed to dot-dot-dot are ignored}
}
\description{
Prints the mass spectrometry data encapsulated by the \code{msDat} object
}
