% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankEN.R
\name{print.rankEN}
\alias{print.rankEN}
\title{Basic information for class \code{rankEN}}
\usage{
\method{print}{rankEN}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{rankEN}}}

\item{...}{Arguments passed to dot-dot-dot are ignored}
}
\description{
Displays the data dimensions used to fit the elastic net model
}

