% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankLasso.R
\name{rankLasso}
\alias{rankLasso}
\title{Ranks compounds using the Lasso path}
\usage{
rankLasso(msDat, activity, actAve, region, useAve, useFirst, numRank)
}
\arguments{
\item{msDat}{The \emph{mass spectrometry} data}

\item{actDat}{The bioactivity data}
}
\description{
Returns a vector containing the compounds in the order in which they first
enter the Lasso model
}

