% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinMS.R
\name{summary.binMS}
\alias{summary.binMS}
\title{Overview of the binning process}
\usage{
\method{summary}{binMS}(binObj)
}
\arguments{
\item{binObj}{An object of class \code{\link{binMS}}}
}
\description{
Prints a text description of the binning process.  Displays arguments passed
to the \code{binMS} routine, how many m/z levels were chosen for each
criterion, how many candidate compounds were chosen overall, and how many
candidate compounds were obtained after consolidation.
}

