% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FilterMS.R
\name{summary.filterMS}
\alias{summary.filterMS}
\title{Overview of the filtering process}
\usage{
\method{summary}{filterMS}(filtObj)
}
\description{
Prints a text description of the filtering process.  Displays arguments
chosen for the \code{filterMS} constructor and how many candidate compounds
were chosen for each criterion, as well as overall.
}

