% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankEN.R
\name{summary.rankEN}
\alias{summary.rankEN}
\title{Overview of the elastic net selection process}
\usage{
\method{summary}{rankEN}(rankEN_obj, max_comp_print = 20L)
}
\arguments{
\item{rankEN_obj}{An object of class \code{rankEN}.}

\item{max_comp_print}{A numeric value >= 1 specifying the maximum number of
  compounds to print}
}
\description{
Prints a description of the elastic net variable selection process.  Includes
the dimensions used to fit the elastic net model, the fraction names for the
mass spectrometry and the bioactivity data in the region of interest, the
parameter specifications for the model, and a table with the identifying
information of the candidate compounds produced by the model fit.
}

