% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioact}
\alias{bioact}
\title{Bioactivity data}
\format{A \code{list} containing relative growth inhibition of bioactivity
    levels for the bacteria and virus strains listed below.  Each of the
    following elements in the \code{list} is a \code{data.frame} with 3 rows
    and 44 columns (with the exception of \emph{F. graminearum} which has 2
    rows).  The rows in each \code{data.frame} correspond to replications of
    the data collection process, while the columns correspond to relative
    growth inhibition bioactivity levels when subject to peptide libraries
    across fractions 1-43 and fraction 47.

\describe{

    \item{ec}{\emph{E. coli}}

    \item{bc}{\emph{S. aureus}}

    \item{pc}{\emph{K. pneumoniae}}

    \item{oc}{\emph{A. baumannii}}

    \item{ab}{\emph{A. baumannii}}

    \item{pa}{\emph{P. aeruginosa}}

    \item{fg}{\emph{F. graminearum}}

}}
\usage{
data(bioact)
}
\description{
The relative relative growth inhibition of bioactivity levels for the
    bacteria and virus strains studies in Kirkpatrick et al. (2016).
}
\keyword{datasets}
