% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankLasso.R
\name{conv_bioact}
\alias{conv_bioact}
\title{Convert bioactivity data}
\usage{
conv_bioact(bioact, bioIdx, useAve)
}
\arguments{
\item{bioact}{Either a non-list vector, a matrix, or a data frame providing
  bioactivity data.  If a non-list vector, then it is assumed that each entry
  corresponds to a particular fraction.  If the data is 2-dimensional, then
  it is assumed that each column corresponds to a particular fraction, and
  that each row corresponds to a particular bioactivity replicate.}

\item{useAve}{A logical value specifying whether or not to average replicate
bioactivity observations.  Ignored if only one bioactivity observation is
provided.}

\item{msIdx}{An integer vector providing the indices (and hence fractions)
for the bioactivity data which are to be used in the Lasso model}
}
\description{
Creates the response vector for use by the \code{lars} function. This
involves transposing the data and reducing the data to the fractions of
interest.  If replicates are to be used, then the fractions containing the
replicates are concatenated together.  If the replicates are to be averaged,
then the data within a fraction is averaged before concatenating the values.
}
\details{
It is important to define how the ordering of the replicates is performed so
that this behavior matches that of \code{\link{conv_ms}}.  The current
implementation places replicates within a fraction consecutively, so that for
example, the data may look like: fraction 1 replicate 1, fraction 1 replicate
2, fraction 1 replicate 3, fraction 2 replicate 1, fraction 2 replicate 2,
fraction 2 replicate 3, fraction 3 replicate 1, ...
}

