% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankLasso.R
\name{getCmpIdx}
\alias{getCmpIdx}
\title{Extract compound ordering from \code{lars} fit}
\usage{
getCmpIdx(lars_fit)
}
\arguments{
\item{lars_fit}{An object of class \code{lars}}
}
\value{
An integer vector indexing (some of the) compounds by the order in
  which they first enter the Lasso model, from first to last.
}
\description{
Returns an integer vector indexing (some of the) compounds by the order in
which they first enter the Lasso model, from first to last.
}

