% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MsDat.R
\name{msDat}
\alias{msDat}
\title{Constructor for class \code{msDat}}
\usage{
msDat(mass_spec, mtoz, charge, ms_inten = NULL)
}
\arguments{
\item{mass_spec}{Either a \code{matrix} or \code{data.frame}.  This object
  must contain mass spectrometry abundances, and may optionally contain
  mass-to-charge values, charge state information, or additional extraneous
  variables.  The mass spectrometry data is expected to be in a form with
  each column corresponding to a variable and each row corresponding to a
  mass-to-charge level.

  For example, suppose that a collection of mass spectrometry intensity
  observations has provided data for 50 fractions across 20,000
  mass-to-charge values.  Then the input for \code{mass_spec} should be a
  \code{matrix} or \code{data.frame} with 20,000 rows and 50 or more columns.
  The additional columns beyond the 50 containing the mass spectrometry
  intensities can be the mass-to-charge data, the charge data, or other
  extraneous variables (which will be ignored when constructing the
  \code{msDat} object).}

\item{mtoz}{A vector of either length 1 or length equal to the number of
  mass-to-charge values for which mass spectrometry data was collected, and
  which helps identify the mass-to-charge values for this data in one of
  several ways.

  One way to provide the information is to provide a numeric vector where
  each entry provides the mass-to-charge value for a corresponding row of
  mass spectrometry data.  Then the \code{k}-th entry of the vector would
  provide the mass-to-charge value for the \code{k}-th row of the mass
  spectrometry data.

  A second way is to provide a single number which specifies the column
  number in the \code{matrix} or \code{data.frame} provided as the argument
  for the \code{mass_spec} parameter such that this column contains the
  mass-to-charge information.

  A third way is provide a single character string which provides the column
  name in the \code{matrix} or \code{data.frame} provided as the argument for
  the \code{mass_spec} parameter such that this column contains the
  mass-to-charge information.  Partial matching is supported.}

\item{charge}{The information for the \code{charge} parameter can be provided
  in the same manner as for the mass-to-charge values.}

\item{ms_inten}{Either \code{NULL} or a vector either of mode character or
  mode numeric specifying which of the variables in the argument to
  \code{mass_spec} are to be retained as the mass spectrometry intensity
  data.  If \code{NULL}, then it is taken to mean that the entirety of the
  data in \code{mass_spec}, after removing variables in the data that are
  specified as arguments, is the mass spectrometry intensity data.  If it is
  a numeric vector, then the entries should provide the indices for the
  region of interest in the mass spectrometry data in the argument for
  \code{msObj}. If it is a character vector, then the entries should uniquely
  specify the region of interest through partial string matching.}
}
\value{
Returns an object of class \code{msDat}.  This class is a \code{list}
  with elements described below.  The class is equipped with a \code{print}
  and \code{extractMS} function.

  \describe{

  \item{\code{ms}}{ A \code{matrix} containing mass spectrometry intensity
  readings. Each column provides the mass spectrometry values for a given
  fraction, and each row provides the mass spectrometry values for a given
  mass-to-charge ratio value across the fractions. }

  \item{\code{mtoz}}{ A vector with length equal to the number of
  mass-to-charge values provided in the mass spectrometry data, such that the
  \code{k}-th entry in the vector provides the mass-to-charge value for the
  \code{k}-th row of mass spectrometry data }

  \item{\code{chg}}{ A vector with length equal to the number of
  mass-to-charge values provided in the mass spectrometry data, such that the
  \code{k}-th entry in the vector provides the charge information for the
  \code{k}-th row of mass spectrometry data }

  }
}
\description{
Creates a data structure encapsulating the mass spectrometry intensity
readings as well as identifying information
}
\details{
Since the mass spectrometry data could conceivably be available to
  the researcher in a variety forms, this function attempts to provide a
  uniform data structure for encapsulating this information.  It is the
  fundamental data structure containing the mass spectrometry data used
  internally by the \code{binMS} and \code{filterMS} routines.  The external
  interface for \code{msDat} is provided to the user so that specifying the
  mass spectrometry information can be made in a distinct step from
  performing statistical analyses, which it is hoped makes the downstream
  analysis routines simpler and more intuitive to use.
}

