% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BinMS.R
\name{print.binMS}
\alias{print.binMS}
\title{Print routine for class \code{binMS}}
\usage{
\method{print}{binMS}(binObj)
}
\arguments{
\item{binObj}{An object of class \code{\link{binMS}}}
}
\description{
More details *******
}

