% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMS.R
\name{print.filterMS}
\alias{print.filterMS}
\title{Basic information for class \code{filterMS}}
\usage{
\method{print}{filterMS}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{filterMS}}}

\item{...}{Arguments passed to dot-dot-dot are ignored}
}
\description{
Displays the number of candidate compounds left in the data after filtering
}

