% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MsDat.R
\name{print.msDat}
\alias{print.msDat}
\title{Basic information for class \code{msDat}}
\usage{
\method{print}{msDat}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{msDat}}}

\item{...}{Arguments passed to dot-dot-dot are ignored}
}
\description{
Displays the number of candidate compounds left in the data
}

