% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankEN.R
\name{print.rankEN}
\alias{print.rankEN}
\title{Basic information for class \code{rankEN}}
\usage{
\method{print}{rankEN}(rankEN_obj)
}
\description{
Displays the dimensions used to fit the elastic net model
}

