% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RankLasso.R
\name{rankLasso}
\alias{rankLasso}
\title{Ranks compounds using the Lasso path}
\usage{
rankLasso(ms, act, msIden, region, useAve = TRUE, byName = TRUE)
}
\arguments{
\item{ms}{The mass spectrometry data}

\item{act}{The bioactivity data}
}
\description{
Returns a vector containing the compounds in the order in which they first
enter the Lasso model
}

