% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterMS.R
\name{summary.filterMS}
\alias{summary.filterMS}
\title{Overview of the filtering process}
\usage{
\method{summary}{filterMS}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{filterMS}}}

\item{...}{Arguments passed to dot-dot-dot are ignored}
}
\description{
Prints a description of the filtering process.  Displays arguments chosen for
the \code{filterMS} constructor, how many candidate compounds were chosen for
each criterion, and how many candidate compounds were chosen overall.
}
